+++ src/core/codegen/plugins/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Codegen plugins barrel export
+ */
+
+export { GENERATED_FILE_HEADER, pluginContent } from '../plugin'
+export { typedDocumentStringPlugin } from './typed-document-string'
