+++ src/core/scanning/documents.ts
@@ -3,7 +3,6 @@
  * Framework-agnostic GraphQL client document scanning
  */
 
-import type { CoreExternalService } from '../types/config'
 import type { ScanContext, ScanResult } from '../types/scanning'
 import { relative } from 'pathe'
 import { GRAPHQL_GLOB_PATTERN } from '../constants'
@@ -15,8 +14,8 @@ const REGEX_SPECIAL_CHARS_RE = /[.*+?^${}()|[\]\\]/g
  * Options for scanning documents
  */
 export interface ScanDocumentsOptions {
-  /** External services to exclude from main scan */
-  externalServices?: CoreExternalService[]
+  /** External services to exclude from main scan (only documents field is used) */
+  externalServices?: Array<{ documents?: string[] }>
   /** Client directory relative path */
   clientDirRelative?: string
 }
