+++ src/nitro/codegen/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Nitro codegen barrel export
+ * Preserves the same public API as the previous monolithic codegen.ts
+ */
+
+export { generateClientTypes } from './client-types'
+export { generateExternalTypes } from './external-types'
+export { generateServerTypes } from './server-types'
